backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0

# Web search
duckduckgo-search>=4.0.0

# Image processing
Pillow>=10.0.0
httpx>=0.25.0

# Semantic cache
numpy>=1.24.0
